+++ OOP/src/main/java/homeWork/hw_11_Lists/task_5_Generic/GenericLList.java
@@ -1,20 +1,18 @@
-package homeWork.hw_11_Lists.task_4;
+package homeWork.hw_11_Lists.task_5_Generic;
 
-import homeWork.hw_11_Lists.task_2.List;
-
-public class LList implements List {
-    private Node firstNode;
-    private Node lastNode;
+public class GenericLList<T> implements GenericList<T> {
+    private Node<T> firstNode;
+    private Node<T> lastNode;
     private int size = 0;
 
     @Override
-    public Object get(int index) throws IndexOutOfBoundsException {
+    public T get(int index) throws IndexOutOfBoundsException {
         if (size == 0 || index >= size || index < 0) {
             throw new IndexOutOfBoundsException();
         } else if (index == 0) {
             return firstNode.item;
         }
-        Node result = firstNode;
+        Node<T> result = firstNode;
 
         for (int i = 1; i < size; i++) {
             result = result.next;
@@ -27,17 +25,17 @@ public class LList implements List {
     }
 
     @Override
-    public boolean add(Object item) {
+    public boolean add(T item) {
         if (size == 0) {
-            firstNode = new Node(null, item, null);
+            firstNode = new Node<>(null, item, null);
             lastNode = firstNode;
             size++;
         } else if (size == 1) {
-            lastNode = new Node(firstNode, item, null);
+            lastNode = new Node<>(firstNode, item, null);
             firstNode.next = lastNode;
             size++;
         } else {
-            lastNode.next = new Node(lastNode, item, null);
+            lastNode.next = new Node<>(lastNode, item, null);
             lastNode = lastNode.next;
             size++;
         }
@@ -45,13 +43,13 @@ public class LList implements List {
     }
 
     @Override
-    public boolean addFirst(Object item) {
+    public boolean addFirst(T item) {
         if (size == 0) {
-            firstNode = new Node(null, item, null);
+            firstNode = new Node<>(null, item, null);
             lastNode = firstNode;
             size++;
         } else {
-            firstNode.prev = new Node(null, item, firstNode);
+            firstNode.prev = new Node<>(null, item, firstNode);
             firstNode = firstNode.prev;
             size++;
         }
@@ -59,21 +57,21 @@ public class LList implements List {
     }
 
     @Override
-    public boolean add(int index, Object item) throws IndexOutOfBoundsException {
+    public boolean add(int index, T item) throws IndexOutOfBoundsException {
         if (size == 0 || index >= size || index < 0) {
             throw new IndexOutOfBoundsException();
         }
-        Node oldNode = firstNode;
+        Node<T> oldNode = firstNode;
 
         if (index == 0) {
-            Node newNode = new Node(oldNode.prev, item, oldNode);
+            Node<T> newNode = new Node<>(oldNode.prev, item, oldNode);
             firstNode = newNode;
         } else {
             for (int i = 1; i < size; i++) {
                 oldNode = oldNode.next;
 
                 if (index == i) {
-                    Node newNode = new Node(oldNode.prev, item, oldNode);
+                    Node<T> newNode = new Node<>(oldNode.prev, item, oldNode);
                     oldNode.prev.next = newNode;
                     oldNode.prev = newNode;
                     size++;
@@ -85,12 +83,12 @@ public class LList implements List {
     }
 
     @Override
-    public boolean remove(Object item) {
+    public boolean remove(T item) {
         if (size == 0) {
             return false;
         }
 
-        Node deleteNoda = firstNode;
+        Node<T> deleteNoda = firstNode;
 
         if (deleteNoda.item == item || deleteNoda.item.equals(item)) {
             firstNode = firstNode.next;
@@ -123,8 +121,8 @@ public class LList implements List {
     }
 
     @Override
-    public Object remove(int index) throws IndexOutOfBoundsException {
-        Object result = null;
+    public T remove(int index) throws IndexOutOfBoundsException {
+        T result = null;
 
         if (size == 0 || index >= size || index < 0) {
             throw new IndexOutOfBoundsException();
@@ -137,7 +135,7 @@ public class LList implements List {
             lastNode = lastNode.prev;
             lastNode.next = null;
         } else {
-            Node deleteNoda = firstNode;
+            Node<T> deleteNoda = firstNode;
 
             for (int i = 1; i < size; i++) {
                 deleteNoda = deleteNoda.next;
@@ -161,8 +159,8 @@ public class LList implements List {
     }
 
     @Override
-    public boolean contains(Object item) {
-        Node findingNoda = firstNode;
+    public boolean contains(T item) {
+        Node<T> findingNoda = firstNode;
 
         if (firstNode == null) {
             return false;
@@ -190,12 +188,12 @@ public class LList implements List {
         return size == 0;
     }
 
-    private static class Node {
-        Object item;
-        Node next;
-        Node prev;
+    private static class Node<T> {
+        T item;
+        Node<T> next;
+        Node<T> prev;
 
-        Node(Node prev, Object element, Node next) {
+        Node(Node<T> prev, T element, Node<T> next) {
             this.item = element;
             this.next = next;
             this.prev = prev;
