+++ OOP/src/main/java/homeWork/hw_11_Lists/task_5_Generic/GenericList.java
@@ -1,6 +1,6 @@
-package homeWork.hw_11_Lists.task_2;
+package homeWork.hw_11_Lists.task_5_Generic;
 
-public interface List {
+public interface GenericList<T> {
 
     /**
      * return object by index
@@ -8,7 +8,7 @@ public interface List {
      * @param index number of item into list
      * @return item by index number
      */
-    Object get(int index) throws IndexOutOfBoundsException;
+    T get(int index) throws IndexOutOfBoundsException;
 
     /**
      * add object to list
@@ -16,7 +16,7 @@ public interface List {
      * @param item which adding to list
      * @return true if operation was successfully
      */
-    boolean add(Object item);
+    boolean add(T item);
 
     /**
      * adding object to begin of list
@@ -24,7 +24,7 @@ public interface List {
      * @param item which adding to list
      * @return true if operation was successfully
      */
-    boolean addFirst(Object item);
+    boolean addFirst(T item);
 
     /**
      * adding an object to the list by index number and move all elements after it 1 cell forward
@@ -34,7 +34,7 @@ public interface List {
      * @return true if operation was successfully
      * @throws IndexOutOfBoundsException if index more than, equals to size of list or less than zero
      */
-    boolean add(int index, Object item) throws IndexOutOfBoundsException;
+    boolean add(int index, T item) throws IndexOutOfBoundsException;
 
     /**
      * delete the element from the list
@@ -42,7 +42,7 @@ public interface List {
      * @param item the element which will be delete from the list
      * @return true if the element was deleted
      */
-    boolean remove(Object item);
+    boolean remove(T item);
 
     /**
      * delete the element from the list by index
@@ -50,7 +50,7 @@ public interface List {
      * @param index number of the element which will be delete from the list
      * @return the element which was deleted
      */
-    Object remove(int index) throws IndexOutOfBoundsException;
+    T remove(int index) throws IndexOutOfBoundsException;
 
     /**
      * deletes all elements from the list and return the list to its initial state
@@ -63,7 +63,7 @@ public interface List {
      * @param item element with which compare list`s elements
      * @return true if element is into the list
      */
-    boolean contains(Object item);
+    boolean contains(T item);
 
     /**
      * return size of the list
